Update to use event handling and mutex around engine
The engine isn't totally threadsafe - but it's pretty lame to construct
two instances.

Use a shared instance and use a mutex when calling into the engine.

Also, move the GameThread code to a ChessEngine worker class, and use a
vanilla QThread that owns the ChessEngine to process events from the
main thread.