Dynamically allocate the ChessEngine
Since it gets moved to the other thread, we want to delete the object
when the other thread exits.

Technically, this isn't needed, because we wait for the
ChessEngineThread to exit before we exit the main thread where the
ChessEngine lives.

But let's do it this way to be more general.