Remove notifier func from Config
The notifier function doesn't really belong on the ChessGame::Config
class.

It's only used by the Chess engine thread - and currently it's
getting set to an uninitialized value right now.

Instead of wrangling the problems with that, move the std::function to
a separate member in `ChessGame`.

Whenever we update send a `ChessGame::Config` to the engine thread,
also send a new notifier func with it. Also add a new atomic config
Id field that we update whenver we update the config so that we can
interrupt the thread to update the new config.

This seems to solves almost all the problems I introduced.
Now I just need to investigate why the config update seems to be
delayed.