Fix overlay generous castle penalty causing piece sacrifice
- Add a test where the engine was sacrificing a bishop to
undermine its opponent's castle position
- Reduce the position multiplier to fix anoter test that fails
when the castle penalty is reduced. Apparently the high castle
penalty was compensating for another problem there. The test in
questio was ""Promoted pawn is promoted to highest value piece
even when capturing", and was moving another piece instead of
capturing a bishop to promote to a (to-be-sacrificed) queen.