Fix logger interface
The standard logger is returning a reference to a local static variable, but
the `make_standard_logger` factory function takes parameters.

Change the return type to a `unique_ptr` and update the users.

Also, use `static constexpr int` in place of `constexpr in a few places,
and make a custom logger that synchronizes with Qt's logger so the output
isn't mixed up.