Fix CMake build
I don't understand this completely, but it works now. Here, I'm making
the WebIDL helper script output an "hpp" file instead of "cpp" so that
I can put it inside the dependencies of the target and CMake won't try
to link into into the target (which would create duplicate symbols).

I also add a LINK_DEPENDS option to make sure it re-links so when the js
or the glue.js is updated, we re-link.