Map to_string (Coord) => asString(Coord)
The name `myNamespace::toString(MyType) -> string` has an unfortunate interaction with the doctest library - you can't use that type in an assertion because it tries to use type to convert the value to a different string type `doctest::String` - which is not constructible from a std::string. So if you use a std::string, any assertions using that type will cause the build to fail.

Workaround it by using the name `asString()` instead of `toString()`