Use a private tag to use make_unique instead of new operator
The allocator tries to only be allocated in the heap as a pointer, since it
doesn't really make sense for it to exist on the stack. So it makes the default
constructor private, and has a factory function to create it.

This creates problems for calling std::make_unique though, which doesn't have
permission to call that method to construct the class. We could try to
make it a friend, but I'm not sure what the syntax for that would be.

An easier way to encapsulate the class is to have a public constructor taking
as an argument a struct that is private to the class.

This way, functions outside the class still can't call that constructor because
they can't construct the private tag, but inside the factory method we can call
std::make_unique just fine.