Refactor the command line UI
Refactor the command line UI to use one function that
returns one of any number of commands, and then a separate
function that handles each of those commands.

This uses a variant. May want to use the overloaded
pattern for this.

This simplifies the handling and ownership significantly.
I also change the game and ConsoleGameStatusUpdater ownership
relationship, which was unclear and bug-prone. Instead,
the Game object is inside a shared pointer inside a ConsoleGame
class. To update the Game object, we just assign to that pointer,
and there is no trickery with references.

The code is a little larger, but it is simpler.