Remove No_En_Passant_Coord
`No_En_Passant_Coord` is the `First_Coord` (zero), but is
just a special case.

It's better to use `optional<Coord>` to differentiate when
we have a en passant state and not - so it's less easy to mix
those states up.

Also, add a `clearEnPassantTargets()` function for the case
when we need to clear the state explicitly.