Do out of line destructors for some classes
According to Clang coding standards

https://llvm.org/docs/CodingStandards.html#provide-a-virtual-method-anchor-for-classes-in-headers

There can be a problem where a class gets emitted into each translation
unit if it has virtual methods and not any out-of-line virtual methods.

To fix it, put the destructor out of line for those classes that
don't have any.

Also, fixup logging for QML version.